requirements.txt
requests==2.24.0
python_nmap==0.6.1
pyhunter==1.7
beautifulsoup4==4.9.3
python-nmap==0.6.1
Pillow==8.0.1
